Use django.db.models.JSONField
