+++ saleor/core/db/fields.py
@@ -1,7 +1,7 @@
 import json
 from typing import Callable
 
-from django.contrib.postgres.fields.jsonb import JSONField
+from django.db.models import JSONField
 
 
 class SanitizedJSONField(JSONField):
