+++ saleor/shipping/models.py
@@ -1,7 +1,6 @@
 from typing import TYPE_CHECKING, Union
 
 from django.conf import settings
-from django.contrib.postgres.fields import JSONField
 from django.db import models
 from django.db.models import Q
 from django_countries.fields import CountryField
@@ -11,7 +10,6 @@ from measurement.measures import Weight
 from prices import Money, MoneyRange
 
 from ..core.permissions import ShippingPermissions
-from ..core.utils.json_serializer import CustomJsonEncoder
 from ..core.utils.translations import TranslationProxy
 from ..core.weight import (
     WeightUnits,
@@ -169,8 +167,6 @@ class ShippingMethod(models.Model):
     maximum_order_weight = MeasurementField(
         measurement=Weight, unit_choices=WeightUnits.CHOICES, blank=True, null=True
     )
-
-    meta = JSONField(blank=True, default=dict, encoder=CustomJsonEncoder)
 
     objects = ShippingMethodQueryset.as_manager()
     translated = TranslationProxy()
